# Query: Fixes query performance regression and adds mocked query benchmark