--- a/Microsoft.Azure.Cosmos/src/Pagination/Page.cs
+++ b/Microsoft.Azure.Cosmos/src/Pagination/Page.cs
@@ -29,6 +29,9 @@ namespace Microsoft.Azure.Cosmos.Pagination
             this.ActivityId = activityId;
             this.State = state;
 
+#if DEBUG
+            // Only do the additional header validation on debug.
+            // This causes a significant impact to performance and is not necessary for release scenarios.
             if (additionalHeaders != null)
             {
                 foreach (string key in additionalHeaders.Keys)
@@ -39,6 +42,7 @@ namespace Microsoft.Azure.Cosmos.Pagination
                     }
                 }
             }
+#endif
 
             this.AdditionalHeaders = additionalHeaders == null ? EmptyDictionary : additionalHeaders.ToImmutableDictionary();
         }