@kirankumarkolli 2021-08-19
Is the header validation to cover for scenarios where ContinuationToken is not supported?
(RU, ActivityId are even applicable for point operation as well)