@j82w 2021-08-19
> Is the header validation to cover for scenarios where ContinuationToken is not supported?
> (RU, ActivityId are even applicable for point operation as well)

The header validation is make sure that you are not putting a known header value like request charge into the additional header dictionary. This might happen new headers are added to the base class, but should never happen in production.