@j82w 2021-09-10
Benchmark results before:
    "Top10PercentAverageRps": 4829,
    "Top20PercentAverageRps": 4821,
    "Top30PercentAverageRps": 4812,
    "Top40PercentAverageRps": 4805,
    "Top50PercentAverageRps": 4798,
    "Top60PercentAverageRps": 4790,
    "Top70PercentAverageRps": 4779,
    "Top80PercentAverageRps": 4766,
    "Top90PercentAverageRps": 4751,
    "Top95PercentAverageRps": 4737,
    "Top99PercentAverageRps": 4700,
    "Top50PercentLatencyInMs": 2.4613,
    "Top75PercentLatencyInMs": 3.1592,
    "Top90PercentLatencyInMs": 6.6559,
    "Top95PercentLatencyInMs": 10.9384,
    "Top98PercentLatencyInMs": 17.196569999999973,
    "Top99PercentLatencyInMs": 21.84066833333331,
    "MaxLatencyInMs": 209.0017,
    "AverageRps": 4669,

After:
    "Top10PercentAverageRps": 4992,
    "Top20PercentAverageRps": 4981,
    "Top30PercentAverageRps": 4972,
    "Top40PercentAverageRps": 4965,
    "Top50PercentAverageRps": 4959,
    "Top60PercentAverageRps": 4951,
    "Top70PercentAverageRps": 4941,
    "Top80PercentAverageRps": 4928,
    "Top90PercentAverageRps": 4915,
    "Top95PercentAverageRps": 4906,
    "Top99PercentAverageRps": 4880,
    "Top50PercentLatencyInMs": 2.3936,
    "Top75PercentLatencyInMs": 3.018,
    "Top90PercentLatencyInMs": 6.4072,
    "Top95PercentLatencyInMs": 10.6983,
    "Top98PercentLatencyInMs": 16.91974999999986,
    "Top99PercentLatencyInMs": 21.50516833333331,
    "MaxLatencyInMs": 200.9202,
    "AverageRps": 4861,